Have collections with different element types use the same sequence of collection sizes
The size distribution was previously sharing the same random generator that was used to generate the elements.  Different element types like String used the random generator more times resulting in a different sequence of collection sizes compared to other element types.  To fix this, a new random generator is created for the size distribution so that the sequence of generated sizes is always the same regardless of how often the original random generator is used for generating element values.